backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
python-jose==3.4.0
passlib[bcrypt]==1.7.4
pydantic==2.10.6
email-validator==2.2.0
python-multipart==0.0.9
numpy==2.0.2
sentence-transformers==3.0.1

# File processing
PyMuPDF==1.24.10
python-docx==1.1.2
pytesseract==0.3.13
Pillow==10.4.0

# NLP
spacy==3.8.4
